Move REST API code in its own submodule
The whole code is in gnocchi.rest, making it imported as soon as anything in
gnocchi.rest is imported. This might be needed when just importing e.g.
gnocchi.rest.app to get access to the options.
