+++ gnocchi/rest/api-paste.ini
@@ -27,11 +27,11 @@ pipeline = http_proxy_to_wsgi gnocchiversions
 
 [app:gnocchiversions]
 paste.app_factory = gnocchi.rest.app:app_factory
-root = gnocchi.rest.VersionsController
+root = gnocchi.rest.api.VersionsController
 
 [app:gnocchiv1]
 paste.app_factory = gnocchi.rest.app:app_factory
-root = gnocchi.rest.V1Controller
+root = gnocchi.rest.api.V1Controller
 
 [filter:keystone_authtoken]
 use = egg:keystonemiddleware#auth_token
